pyproject.toml
[project]
name = "serenityai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "ollama>=0.5.3",
    "pydantic>=2.11.7",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "twilio>=9.7.2",
    "uvicorn>=0.35.0",
]
